Fix circlular dependency by moving EthWatcher into it's own package
