+++ Makefile
@@ -56,7 +56,7 @@ test-go-parallel:
 
 .PHONY: test-go-serial
 test-go-serial:
-	go test ./ethereum ./zeroex/ordervalidator ./zeroex/orderwatch -race -timeout 30s -p=1 --serial
+	go test ./ethereum/ethwatch ./zeroex/ordervalidator ./zeroex/orderwatch -race -timeout 30s -p=1 --serial
 
 
 .PHONY: test-integration
