+++ ratelimit/rate_limiter.go
@@ -3,6 +3,7 @@ package ratelimit
 import (
 	"context"
 	"errors"
+	"fmt"
 	"sync"
 	"time"
 
@@ -179,6 +180,7 @@ func getUTCMidnightOfDate(date time.Time) time.Time {
 
 // Wait blocks until the rateLimiter allows for another request to be sent
 func (r *RateLimiter) Wait(ctx context.Context) error {
+	fmt.Println("r.twentyFourHourLimiter", r.twentyFourHourLimiter)
 	if err := r.twentyFourHourLimiter.Wait(ctx); err != nil {
 		return err
 	}
